+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/at/JavaScript/index-non-numeric-argument-tointeger-invalid.js
@@ -1,0 +1,24 @@
+// Copyright (C) 2020 Rick Waldron. All rights reserved.
+// This code is governed by the BSD license found in the LICENSE file.
+/*---
+esid: sec-array.prototype.at
+description: >
+  Property type and descriptor.
+info: |
+  Array.prototype.at( index )
+
+  Let relativeIndex be ? ToInteger(index).
+
+features: [Array.prototype.at]
+---*/
+
+console.log(typeof Array.prototype.at === 'function');
+
+let a = [0, 1, 2, 3];
+
+try {
+  a.at(Symbol());
+  console.log(false);
+} catch (e) {
+  console.log(e instanceof TypeError);
+}